Add CMake build
This adds a CMake build with support for Linux, MacOS, and Windows. Mostly copied from https://github.com/Mendeley/breakpad/blob/master/mendeley/CMakeLists.txt with some modifications.

Some binaries are also removed that we don't care about for PyTorch, and the linux-syscall-support folder is added as a proper submodule.